split certificate issuance into separate module
this was bolted onto challenge solely because that already had access to
the config. while already modifying this code, the default_backend
argument was removed as per
<https://cryptography.io/en/latest/faq/#faq-missing-backend>.
