+++ serles/challenge.py
@@ -6,14 +6,11 @@ import base64
 import requests
 import ipaddress
 import jwcrypto.jwk  # fedora package: python3-jwcrypto.noarch
-import jwcrypto.jws
 import dns.resolver
 
 from datetime import datetime, timezone
 
 from cryptography import x509  # python3-cryptography.x86_64
-from cryptography.hazmat.backends import default_backend as x509_backend
-from cryptography.hazmat.primitives import serialization
 
 from .utils import get_ptr, ip_in_ranges, normalize, ber_parse
 from .configloader import get_config
@@ -235,7 +232,7 @@ def alpn_challenge(challenge):  # RFC 8737 §3
                 return "unauthorized", f"could not negotiate {ALPN_PROTOCOL!r}"
 
             cert = x509.load_der_x509_certificate(
-                ssock.getpeercert(binary_form=True), x509_backend()
+                ssock.getpeercert(binary_form=True)
             )
     except (socket.error, ssl.SSLError, ValueError) as e:
         return "connection", str(e)
@@ -261,80 +258,6 @@ def alpn_challenge(challenge):  # RFC 8737 §3
         return "incorrectResponse", "key authorization hashes don't match"
 
     return None, None  # no error occurred :)
-
-
-def check_csr_and_return_cert(csr_der, order):
-    """ validate CSR and pass to backend
-
-    Checks that the CSR only contains domains from previously validated
-    challenges and get a signed certificate from the backend.
-
-    Args:
-        csr_der (bytes): client's CSR in DER encoding
-        order (Order): the order object that the CSR belongs to
-
-    Returns:
-        bytes: the signed certificate and chain in PEM format.
-
-    Raises:
-        ACMEError: CSR was rejected (by us) or the backend refused to sign it.
-    """
-    csr = x509.load_der_x509_csr(csr_der, x509_backend())
-    try:
-        san = csr.extensions.get_extension_for_oid(
-            x509.oid.ExtensionOID.SUBJECT_ALTERNATIVE_NAME
-        ).value
-        san_dns = san.get_values_for_type(x509.DNSName)
-        san_ips = san.get_values_for_type(x509.IPAddress)
-        alt_names = san_dns + san_ips
-    except x509.extensions.ExtensionNotFound as e:
-        san_dns = []
-        san_ips = []
-        alt_names = []
-
-    try:
-        common_name = csr.subject.get_attributes_for_oid(x509.oid.NameOID.COMMON_NAME)[
-            0
-        ].value
-    except IndexError:
-        if not alt_names:
-            raise ACMEError("no identifiers in CSR", 400, "badCSR")
-        common_name = alt_names[0]
-
-    if not common_name in alt_names:  # chrome ignores CN, so write CN to SAN
-        alt_names.insert(0, common_name)
-        try:
-            ip = ipaddress.ip_address(common_name)
-            san_ips.insert(0, ip)
-        except ValueError:
-            san_dns.insert(0, common_name)
-
-    # since we pass the CN and SANs to the backend, make sure the client only
-    # specified those that we verified before (and that types match):
-    order_identifiers = {(unstringify_ip(ident), ident.type) for ident in order.identifiers}
-    csr_identifiers = {
-        *((ident, IdentifierTypes.dns) for ident in san_dns),
-        *((ident, IdentifierTypes.ip) for ident in san_ips),
-    }
-
-    if order_identifiers != csr_identifiers:
-        raise ACMEError(f"{order_identifiers} != {csr_identifiers}", 400, "badCSR")
-
-    csr_pem = csr.public_bytes(serialization.Encoding.PEM)
-    email = order.account.contact
-    subject_dn = csr.subject.rfc4514_string()
-    if config["forceTemplateDN"] or not subject_dn:
-        subject_dn = config["subjectNameTemplate"].format(SAN=alt_names, MAIL=email)
-
-    certificate, error = backend.sign(csr_pem, subject_dn, alt_names, email)
-
-    if error:
-        raise ACMEError(error, 400, "badCSR")
-
-    if type(certificate) == str:
-        certificate = certificate.encode("utf-8")
-
-    return certificate
 
 
 def key_authorization(challenge):
@@ -381,21 +304,3 @@ def additional_ip_address_checks(remote_ip, host, is_ipaddress=False):
     if not is_ipaddress:
         if config["verifyPTR"] and normalize(get_ptr(remote_ip)) != normalize(host):
             return "rejectedIdentifier", f"PTR does not match"
-
-def unstringify_ip(ident):
-    """ convert IP address identifiers to `ipaddress` objects
-
-    cryptography.x509.SubjectAlternativeName.get_values_for_type returns IP
-    addresses as IPv4Address/IPv6Address objects, not strings. This allows us
-    to compare Order identifiers to them.
-
-    Args:
-        ident (Identifier): the identifier to convert based on its type
-
-    Returns:
-        str | IPv4Address | IPv6Address: converted value
-    """
-
-    if ident.type == IdentifierTypes.ip:
-        return ipaddress.ip_address(ident.value)
-    return ident.value
