+++ tests/test_challengefuns.py
@@ -6,7 +6,6 @@ import unittest
 from unittest.mock import Mock
 from unittest.mock import MagicMock
 from unittest.mock import PropertyMock
-import pytest
 from copy import deepcopy
 
 import serles.challenge as main
@@ -680,134 +679,3 @@ class ChallengeFunctionTester(unittest.TestCase):
         ):
             result = main.alpn_challenge(mock_alpn_challenge_brokenip)
             self.assertEqual(result[0], "malformed")
-
-    def test_check_csr_and_return_cert(self):
-        csr_input = open("data_example.test.csr.bin", "rb").read()
-        mock_order = Mock()
-        mock_order.account.contact = None
-        example_inval = Mock()
-        example_inval.value = "example.inval"
-        example_inval.type = IdentifierTypes.dns
-        example_test = Mock()
-        example_test.value = "example.test"
-        example_test.type = IdentifierTypes.dns
-
-        # additional identifiers in CSR:
-        mock_order.identifiers = []
-        self.assertRaisesRegex(
-            main.ACMEError,
-            r"set\(\)",
-            main.check_csr_and_return_cert,
-            csr_input,
-            mock_order,
-        )
-        # identifiers missing in CSR:
-        mock_order.identifiers = [example_inval, example_test]
-        self.assertRaisesRegex(
-            main.ACMEError,
-            "example.inval",
-            main.check_csr_and_return_cert,
-            csr_input,
-            mock_order,
-        )
-        mock_order.identifiers = [example_test]
-
-        result = main.check_csr_and_return_cert(csr_input, mock_order)
-        good = open("data_pkcs7.bin", "rb").read()
-        self.assertEqual(result, good)
-
-        with unittest.mock.patch.object(
-            main.backend, "sign", lambda *x: ("a string, not bytes", None)
-        ):
-            result = main.check_csr_and_return_cert(csr_input, mock_order)
-            self.assertEqual(result, b"a string, not bytes")  # utf-8-encoded bytes
-
-        with unittest.mock.patch.object(
-            main.backend, "sign", lambda *x: (None, "error")
-        ):
-            self.assertRaisesRegex(
-                main.ACMEError,
-                "error",
-                main.check_csr_and_return_cert,
-                csr_input,
-                mock_order,
-            )
-
-    def test_check_csr_and_return_cert_nocn(self):
-        csr_input = open("data_nocn.csr", "rb").read()
-        mock_order = Mock()
-        mock_order.account.contact = None
-        example_inval = Mock()
-        example_inval.value = "example.inval"
-        example_inval.type = IdentifierTypes.dns
-        example_test = Mock()
-        example_test.value = "example.test"
-        example_test.type = IdentifierTypes.dns
-
-        # additional identifiers in CSR:
-        mock_order.identifiers = []
-        self.assertRaisesRegex(
-            main.ACMEError,
-            r"set\(\)",
-            main.check_csr_and_return_cert,
-            csr_input,
-            mock_order,
-        )
-        # identifiers missing in CSR:
-        mock_order.identifiers = [example_inval, example_test]
-        self.assertRaisesRegex(
-            main.ACMEError,
-            "example.inval",
-            main.check_csr_and_return_cert,
-            csr_input,
-            mock_order,
-        )
-        mock_order.identifiers = [example_test]
-
-        result = main.check_csr_and_return_cert(csr_input, mock_order)
-        good = open("data_pkcs7.bin", "rb").read()
-        self.assertEqual(result, good)
-
-        with unittest.mock.patch.object(
-            main.backend, "sign", lambda *x: (None, "error")
-        ):
-            self.assertRaisesRegex(
-                main.ACMEError,
-                "error",
-                main.check_csr_and_return_cert,
-                csr_input,
-                mock_order,
-            )
-
-    def test_check_csr_and_return_cert_ipcn(self):
-        csr_input = open("data_csr_ipcn.der", "rb").read()
-        mock_order = Mock()
-        mock_order.account.contact = None
-        example_ip = Mock()
-        example_ip.value = "2001:DB8::1"
-        example_ip.type = IdentifierTypes.ip
-        mock_order.identifiers = [example_ip]
-
-        result = main.check_csr_and_return_cert(csr_input, mock_order)
-        good = open("data_pkcs7.bin", "rb").read()
-        self.assertEqual(result, good)
-
-    def test_check_csr_and_return_cert_nocnorsan(self):
-        csr_input = open("data_csr_empty.der", "rb").read()
-        mock_order = Mock()
-        mock_order.account.contact = None
-        example_inval = Mock()
-        example_inval.value = "example.inval"
-        example_inval.type = IdentifierTypes.dns
-        example_test = Mock()
-        example_test.value = "example.test"
-        example_test.type = IdentifierTypes.dns
-
-        mock_order.identifiers = []
-        self.assertRaisesRegex(
-            main.ACMEError,
-            r"no identifiers in CSR",
-            main.check_csr_and_return_cert,
-            csr_input,
-            mock_order,
-        )
